[Refactor] Type Enum Class Package Refactoring
