+++ app/src/main/java/com/angdroid/refrigerator_manament/presentation/util/BindingConversions.kt
@@ -1,10 +1,13 @@
 package com.angdroid.refrigerator_manament.presentation.util
 
 import android.widget.ImageView
+import android.widget.TextView
 import androidx.databinding.BindingAdapter
 import coil.load
 import coil.transform.RoundedCornersTransformation
 import com.angdroid.refrigerator_manament.R
+import com.angdroid.refrigerator_manament.presentation.util.types.CategoryType
+import com.angdroid.refrigerator_manament.presentation.util.types.FoodTypeFeatures
 
 
 @BindingAdapter("app:load_remote_coil_corner")
@@ -14,13 +17,16 @@ fun ImageView.loadRemoteCoilCorner(url: String) {
         this.load(url) {
             transformations(RoundedCornersTransformation(12.0F))
         }
-    } else {
-        this.load(R.drawable.temp1)
     }
 }
 
 
 @BindingAdapter("app:load_default_ingredient")
-fun ImageView.loadDefaultIngredient(foodName:String){
+fun ImageView.loadDefaultIngredient(foodName: String) {
     this.setImageResource(FoodTypeFeatures.valueOf(foodName).imageRes)
+}
+
+@BindingAdapter("app:category_text")
+fun TextView.categoryText(categoryId: Int) {
+    text = resources.getText(CategoryType.categoryList[categoryId - 1])
 }
