--- app/src/main/java/com/angdroid/refrigerator_manament/presentation/util/FoodType.kt
@@ -1,9 +1,0 @@
-package com.angdroid.refrigerator_manament.presentation.util
-
-import com.angdroid.refrigerator_manament.R
-
-enum class FoodType {
-     당근, 오이, 무우, 사과, 배, 귤, 계란, 생닭, 우유, 치즈, 새우, 오징어, 고등어
-}
-
-//CARROT, CUCUMBER, MOO, APPLE, PEAR, TANGERINE, EGG, CHICKEN, MILK, CHEESE, SHRIMP,SQUID, MACKEREL
