+++ app/src/main/res/layout/item_category_list.xml
@@ -29,8 +29,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:load_default_ingredient="@{food.name}"
-            tools:src="@drawable/temp1" />
+            app:load_default_ingredient="@{food.name}" />
 
         <TextView
             android:id="@+id/iv_remove"
