+++ app/src/main/res/layout/item_category_title.xml
@@ -19,7 +19,7 @@
             android:id="@+id/tvCategory"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:text="@{titleModel.categoryName}"
+            app:category_text="@{titleModel.category}"
             android:textColor="@color/black"
             android:textSize="16dp"
             app:layout_constraintEnd_toEndOf="parent"
